fix(condo): DOMA-11189 move constants to separate file
