+++ packages/keystone/file/constants.js
@@ -1,0 +1,7 @@
+const ROWS_COUNT_LIMIT = 100_000
+const ROWS_COUNT_LIMIT_EXCEEDED = 'Rows count limit exceeded'
+
+module.exports = {
+    ROWS_COUNT_LIMIT,
+    ROWS_COUNT_LIMIT_EXCEEDED,
+}
